mm: implement WasActive page flag (for improving cleancache)
Since about 2.6.27, the page replacement algorithm maintains
an "active" bit to help decide which pages are most eligible
to reclaim, see http://linux-mm.org/PageReplacementDesign

This "active' information is also useful to cleancache but is lost
by the time that cleancache has the opportunity to preserve the
pageful of data. This patch adds a new page flag "WasActive" to
retain the state. The flag may possibly be useful elsewhere.

It is up to each cleancache backend to utilize the bit as
it desires. The matching patch for zcache is included here
for clarification/discussion purposes, though it will need to
go through GregKH and the staging tree.

The patch resolves issues reported with cleancache which occur
especially during streaming workloads on older processors,
see https://lkml.org/lkml/2011/8/17/351

Change-Id: I560c9a2ca1134ca6247e7037690b4b2c9340a72a
Signed-off-by: Dan Magenheimer <***@oracle.com>